Math_Foundations

This package is a collection of functions and notebooks I built as I worked through the book Foundations of Mathematics: A Preparatory Course.

In essence, this book is a primer on basic mathematical concepts serving as a bridge into college mathematics. It starts with the very basics of arithmetical computation, moving onto basic mathematical functions, geometry and trigonometry. From there it covers more advanced topics such as basic linear algebra, calculus, probability and statistics, and concluding with complex numbers.

The English edition is a mostly AI translation of the fourth edition of the original German text. There are a lot of dad jokes and bad puns in the text, but it is definitely clearly written and explains the concepts well. The very approachable style makes it quite suitable for self-study. I used it as a review in preparation of my study for more advanced mathematical topics.

The code in this package is not meant to be a complete implementation of the book, but rather stuff that I created in order to gain practice and a deeper understanding of the concepts in the book. The code is all in Julia, a language I find extremely suitable for mathematical study and actually fun to code with. I make use of multiple Julia packages, not because they are necessary to implement a specific concept, but rather, in order to have an opportunity to play with and learn about some of their capabilities.

Basic Mathematical Functions

This section provides the documentation for all the functions I wrote to help me work through the book.

Math_Foundations.accrued — Method

function accrued(i::Real,p::Real,c::Int64) -> Float64

i is interest, p is period (1 = one year), c is capital return accrued value

Math_Foundations.accrued_apr — Method

function accrued_apr(i::Real,p::Real,c::Int64) -> Float64

i is interest, p is period (1 = one year), c is capital return accrued value using daily apr for interest

Math_Foundations.expa2x — Method

expa2x(a::Real,x::Real)

exponential function where a > 0 and x is any Real

Math_Foundations.nth_root — Method

nth_root(x,n) → c

Calculate nth root of x handling negative x properly

Math_Foundations.plot_hyperbola — Function

p
plot_hyperbola(a::Float64=1.0, h::Float64=0.0, k::Float64=0.0)

Plot hyperbola with equation y = a/(x-h) + b, where a,h and k are scaling parameters

Math_Foundations.plot_hyperbola_axes_direct — Method

plot_hyperbola_axes_direct(a::Float64, b::Float64)

Plot hyperbola of the form (x²/a²) - (y²/b²) = 1 a and b are the semi-major and semi-minor axes

Math_Foundations.plot_hyperbola_axes_varx — Method

plot_hyperbola_axes_varx(a::Float64, b::Float64)

Plot hyperbola of the form (x²/a²) - (y²/b²) = 1 a and b are the semi-major and semi-minor axes

Math_Foundations.plot_parabola — Method

plot_parabola(p::Polynomial, r3::Union{Vector{Float64},Vector{ComplexF64}}, a₂::Float64, a₁::Float64, a₀::Float64, str::String)

Plot the parabola using the given polynomial and its roots

Math_Foundations.plot_parabola_roots_amrvw — Function

plot_parabola_roots_amrvw(a₂::Float64, a₁::Float64=0.0, a₀::Float64=0.0) -> [ComplexF64, ComplexF64]

AMRVW version of plotting parabola with roots

Math_Foundations.plot_parabola_roots_polynomial — Function

plot_parabola_roots_polynomial(a₂::Float64, a₁::Float64=0.0, a₀::Float64=0.0) -> [ComplexF64, ComplexF64]

Polynomial version of plotting parabola with roots

Math_Foundations.plot_parabola_roots_quadratic — Function

plot_parabola_roots_quadratic(a₂::Float64, a₁::Float64=0.0, a₀::Float64=0.0) -> [ComplexF64, ComplexF64]

Quadratic formula version of plotting parabola with roots

Math_Foundations.triangle_area_perim — Method

triangle_area_perim(a::Float64, b::Float64, c::Float64) -> Float64

Calculate  ρ, area and perimeter of a triangle using Heron's formula

External Documentation

This section provides links to external documentation that replicates topics covered in the book and often provide more details. It is not an exhaustive list, but it does cover many of the important topics. Most of the links are to the LibreTexts project, which is a free online library of mathematics textbooks and resources. The links are grouped by topic.

Basic Computation

Functions

Geometry

Parabola

Hyperbola

Triangle

Trigonometry

Trigonometric Functions