Math_Foundations

This is the mandatory index.md file that is the homepage of the documentation.

Basic Mathematical Functions

Math_Foundations.accrued — Method

function accrued(i::Real,p::Real,c::Int64) -> Float64

i is interest, p is period (1 = one year), c is capital return accrued value

Math_Foundations.accrued_apr — Method

function accrued_apr(i::Real,p::Real,c::Int64) -> Float64

i is interest, p is period (1 = one year), c is capital return accrued value using daily apr for interest

Math_Foundations.expa2x — Method

expa2x(a::Real,x::Real)

exponential function where a > 0 and x is any Real

Math_Foundations.nth_root — Method

nth_root(x,n) → c

Calculate nth root of x handling negative x properly

Math_Foundations.plot_hyperbola — Function

p
plot_hyperbola(a::Float64=1.0, h::Float64=0.0, k::Float64=0.0)

Plot hyperbola with equation y = a/(x-h) + b, where a,h and k are scaling parameters

Math_Foundations.plot_hyperbola_axes_direct — Method

plot_hyperbola_axes_direct(a::Float64, b::Float64)

Plot hyperbola of the form (x²/a²) - (y²/b²) = 1 a and b are the semi-major and semi-minor axes

Math_Foundations.plot_hyperbola_axes_varx — Method

plot_hyperbola_axes_varx(a::Float64, b::Float64)

Plot hyperbola of the form (x²/a²) - (y²/b²) = 1 a and b are the semi-major and semi-minor axes

Math_Foundations.plot_parabola — Method

plot_parabola(p::Polynomial, r3::Union{Vector{Float64},Vector{ComplexF64}}, a₂::Float64, a₁::Float64, a₀::Float64, str::String)

Plot the parabola using the given polynomial and its roots

Math_Foundations.plot_parabola_roots_amrvw — Function

plot_parabola_roots_amrvw(a₂::Float64, a₁::Float64=0.0, a₀::Float64=0.0) -> [ComplexF64, ComplexF64]

AMRVW version of plotting parabola with roots

Math_Foundations.plot_parabola_roots_polynomial — Function

plot_parabola_roots_polynomial(a₂::Float64, a₁::Float64=0.0, a₀::Float64=0.0) -> [ComplexF64, ComplexF64]

Polynomial version of plotting parabola with roots

Math_Foundations.plot_parabola_roots_quadratic — Function

plot_parabola_roots_quadratic(a₂::Float64, a₁::Float64=0.0, a₀::Float64=0.0) -> [ComplexF64, ComplexF64]

Quadratic formula version of plotting parabola with roots

Math_Foundations.triangle_area_perim — Method

triangle_area_perim(a::Float64, b::Float64, c::Float64) -> Float64

Calculate  ρ, area and perimeter of a triangle using Heron's formula

External Documentation

Geometry Hyperbola